Move PathToKey to own file
